requirements.txt
gradio>=4.0
python-dotenv>=1.0
requests>=2.31
pyaudio>=0.2.13
SpeechRecognition>=3.10
soundfile>=0.12
pyttsx3>=2.90
pillow>=10.0
tenacity>=8.2
pytest>=7.4